refactor(journal): fix import paths
(cherry picked from commit bf93d8f3a6b401ecd6b762dca11921a7777a3617)
